Layer changes: Add BaseBuilder for the layer dsl to make it easier to read Remove unused optimizers (non sgd) Remove helpers
