+++ deeplearning4j/deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/api/OptimizationAlgorithm.java
@@ -26,5 +26,5 @@ package org.deeplearning4j.nn.api;
  *
  */
 public enum OptimizationAlgorithm {
-    LINE_GRADIENT_DESCENT, CONJUGATE_GRADIENT, LBFGS, STOCHASTIC_GRADIENT_DESCENT
+    STOCHASTIC_GRADIENT_DESCENT
 }
